requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu
gradio>=4.30
pillow
numpy